Accidentally made it pull git for the current working directory whatever that is, should only do this for PADRE directory :(
Additionally made it so that it retries looking for the PADRE connection string, by default every 0.25 seconds for 10 seconds but can be configured. Suspect this is good enough for most.
